+++ src/game/nlcore/game.ts
@@ -50,6 +50,9 @@ export class Game {
             skipKey: ["Control"],
             skipInterval: 100,
             ratioUpdateInterval: 50,
+            preloadDelay: 100,
+            preloadConcurrency: 5,
+            waitForPreload: false,
         },
         elements: {
             say: {