+++ src/game/player/elements/image/AspectScaleImage.tsx
@@ -1,6 +1,9 @@
 import React, {useEffect, useRef} from "react";
 import {ImgElementProp} from "@core/elements/transition/type";
 import {useRatio} from "@player/provider/ratio";
+import {usePreloaded} from "@player/provider/preloaded";
+import {Image} from "@core/elements/displayable/image";
+import {useGame} from "@core/common/player";
 
 export default function AspectScaleImage(
     {
@@ -18,6 +21,8 @@ export default function AspectScaleImage(
     const imgRef = useRef<HTMLImageElement>(null);
     const {ratio} = useRatio();
     const [width, setWidth] = React.useState<number>(0);
+    const {cacheManager} = usePreloaded();
+    const {game} = useGame();
 
     function updateWidth() {
         const ref = Ref || imgRef;
@@ -29,12 +34,18 @@ export default function AspectScaleImage(
     useEffect(() => {
         updateWidth();
 
+        if (props.src && !cacheManager.has(props.src)) {
+            game.getLiveGame().getGameState()?.logger.warn("AspectScaleImage", "Image not preloaded", props.src);
+        }
+
         return ratio.onUpdate(updateWidth);
     }, [props.src]);
 
     useEffect(() => {
         updateWidth();
     }, [props, id]);
+
+    const src: string = props.src ? (cacheManager.get(props.src) || props.src) : Image.DefaultImagePlaceholder;
 
     return (
         <img
@@ -43,6 +54,7 @@ export default function AspectScaleImage(
             onLoad={onLoad}
             width={width}
             alt={props.alt}
+            src={src}
         />
     );
 }